requirements.txt
Adafruit-Blinka==6.18.0
adafruit-circuitpython-busdevice==5.1.1
adafruit-circuitpython-framebuf==1.4.8
adafruit-circuitpython-ina219==3.4.10
adafruit-circuitpython-register==1.9.7
adafruit-circuitpython-ssd1306==2.12.3
Adafruit-PlatformDetect==3.19.2
Adafruit-PureIO==1.1.9
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
build==0.7.0
charset-normalizer==2.0.10
flake8==4.0.1
frozenlist==1.2.0
idna==3.3
importlib-metadata==4.2.0
iniconfig==1.1.1
mccabe==0.6.1
multidict==5.2.0
packaging==21.3
pep517==0.12.0
Pillow==9.0.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyftdi==0.53.3
pyparsing==3.0.6
pyserial==3.5
pytest==6.2.5
pytest-aio==1.4.1
pytest-mock==3.6.1
python-engineio==3.14.2
python-socketio==4.6.1
pyusb==1.2.1
rpi-ws281x==4.3.1
RPi.GPIO==0.7.1
six==1.16.0
sysv-ipc==1.1.0
toml==0.10.2
tomli==2.0.0
typing_extensions==4.0.1
websockets==10.1
yarl==1.7.2
zipp==3.7.0
